批量发送到MQ
